+++ base/index.js
@@ -1,20 +1,23 @@
+const { SORT_IMPORTS_GROUPS } = require("../lib/eslint-plugin-perfectionist");
+
 /**
  * Workaround to allow ESLint to resolve plugins that were installed
  * by an external config, see https://github.com/eslint/eslint/issues/3458.
  */
 require("@rushstack/eslint-patch/modern-module-resolution");
 
-/** @type {import('eslint').ESLint.ConfigData & { parserOptions: import('eslint').ESLint.ConfigData & import('@typescript-eslint/parser').ParserOptions } }  */
+/** @type {import('eslint').ESLint.ConfigData & { parserOptions: import('eslint').ESLint.ConfigData['parserOptions'] & import('@typescript-eslint/parser').ParserOptions } }  */
 module.exports = {
   env: {
     es2022: true,
   },
   extends: [
     "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:@typescript-eslint/recommended-requiring-type-checking",
+    "plugin:@typescript-eslint/recommended-type-checked",
+    "plugin:@typescript-eslint/stylistic-type-checked",
     "plugin:import/recommended",
     "plugin:import/typescript",
+    "plugin:perfectionist/recommended-natural",
     "plugin:sonarjs/recommended",
   ],
   overrides: [
@@ -34,9 +37,10 @@ module.exports = {
     // find the tsconfig.json nearest each source file
     project: true,
   },
-  plugins: ["@typescript-eslint", "sonarjs", "sort-keys-plus"],
+  plugins: ["@typescript-eslint", "sonarjs"],
   rules: {
     // @typescript-eslint: https://github.com/typescript-eslint/typescript-eslint/tree/main/packages/eslint-plugin/docs/rules
+    "@typescript-eslint/adjacent-overload-signatures": "off", // disabled due to conflict with eslint-plugin-perfectionist
     "@typescript-eslint/no-floating-promises": ["error", { ignoreVoid: true }],
     "@typescript-eslint/no-misused-promises": [
       "error",
@@ -52,7 +56,7 @@ module.exports = {
         varsIgnorePattern: "^_",
       },
     ],
-    "@typescript-eslint/sort-type-constituents": "error",
+    "@typescript-eslint/sort-type-constituents": "off", // disabled due to conflict with eslint-plugin-perfectionist
 
     // eslint: https://github.com/eslint/eslint/tree/main/lib/rules
     "arrow-body-style": ["error", "as-needed"],
@@ -75,6 +79,8 @@ module.exports = {
     "prefer-const": "error",
     "prefer-rest-params": "error",
     "prefer-template": "error",
+    "sort-imports": "off", // disabled due to conflict with eslint-plugin-perfectionist
+    "sort-keys": "off", // disabled due to conflict with eslint-plugin-perfectionist
 
     // eslint-plugin-import: https://github.com/import-js/eslint-plugin-import/tree/main/docs/rules
     "import/no-cycle": "error",
@@ -86,30 +92,40 @@ module.exports = {
         unusedExports: true,
       },
     ],
-    "import/order": [
-      "error",
-      {
-        alphabetize: {
-          caseInsensitive: true,
-          order: "asc",
-          orderImportKind: "asc",
-        },
-      },
-    ],
+    "import/order": "off", // disabled due to conflict with eslint-plugin-perfectionist
     "import/prefer-default-export": "off",
 
-    /**
-     * Required to fix sort-keys automatically, since this is not done by default.
-     * See: https://github.com/forivall/eslint-plugin-sort-keys-plus
-     */
-    "sort-keys-plus/sort-keys": [
+    // Deactivated as TypeScript provides the same checks as part of standard type checking: https://typescript-eslint.io/linting/troubleshooting/performance-troubleshooting
+    "import/default": "off",
+    "import/named": "off",
+    "import/namespace": "off",
+    "import/no-named-as-default-member": "off",
+
+    // eslint-plugin-perfectionist: https://github.com/azat-io/eslint-plugin-perfectionist
+    "perfectionist/sort-array-includes": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-astro-attributes": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-classes": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-enums": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-exports": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-imports": [
       "error",
-      "asc",
       {
-        caseSensitive: false,
-        natural: true,
+        groups: SORT_IMPORTS_GROUPS,
+        "ignore-case": true,
+        "newlines-between": "ignore",
+        type: "natural",
       },
     ],
+    "perfectionist/sort-interfaces": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-jsx-props": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-maps": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-named-exports": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-named-imports": ["error", { "ignore-alias": true, "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-object-types": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-objects": ["error", { "ignore-case": true, "partition-by-comment": true, type: "natural" }],
+    "perfectionist/sort-svelte-attributes": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-union-types": ["error", { "ignore-case": true, type: "natural" }],
+    "perfectionist/sort-vue-attributes": ["error", { "ignore-case": true, type: "natural" }],
   },
   settings: {
     "import/resolver": {