+++ react/index.js
@@ -1,10 +1,12 @@
+const { SORT_IMPORTS_GROUPS } = require("../lib/eslint-plugin-perfectionist");
+
 /**
  * Workaround to allow ESLint to resolve plugins that were installed
  * by an external config, see https://github.com/eslint/eslint/issues/3458.
  */
 require("@rushstack/eslint-patch/modern-module-resolution");
 
-/** @type {import('eslint').ESLint.ConfigData & { parserOptions: import('eslint').ESLint.ConfigData & import('@typescript-eslint/parser').ParserOptions } }  */
+/** @type {import('eslint').ESLint.ConfigData & { parserOptions: import('eslint').ESLint.ConfigData['parserOptions'] & import('@typescript-eslint/parser').ParserOptions } }  */
 module.exports = {
   env: {
     browser: true,
@@ -43,44 +45,45 @@ module.exports = {
     ],
     "@typescript-eslint/consistent-type-definitions": ["error", "type"],
 
-    // eslint-plugin-import: https://github.com/import-js/eslint-plugin-import/tree/main/docs/rules
-    "import/order": [
-      "error",
-      {
-        alphabetize: {
-          caseInsensitive: true,
-          order: "asc",
-          orderImportKind: "asc",
-        },
-        pathGroups: [
-          {
-            group: "external",
-            pattern: "react",
-            position: "before",
-          },
-        ],
-        pathGroupsExcludedImportTypes: ["react"],
-      },
-    ],
-
     // eslint-plugin-react: https://github.com/jsx-eslint/eslint-plugin-react/tree/master/lib/rules
     "react/jsx-pascal-case": "error",
-    "react/jsx-sort-props": [
-      "error",
-      {
-        callbacksLast: true,
-        ignoreCase: true,
-        noSortAlphabetically: false,
-        reservedFirst: true,
-        shorthandFirst: false,
-        shorthandLast: false,
-      },
-    ],
+    "react/jsx-sort-props": "off", // disabled due to conflict with eslint-plugin-perfectionist
     "react/sort-default-props": "error",
 
     // eslint-plugin-react-hooks: https://github.com/facebook/react/blob/main/packages/eslint-plugin-react-hooks/README.md
     "react-hooks/exhaustive-deps": "error",
     "react-hooks/rules-of-hooks": "error",
+
+    // eslint-plugin-perfectionist: https://github.com/azat-io/eslint-plugin-perfectionist
+    "perfectionist/sort-imports": [
+      "error",
+      {
+        "custom-groups": {
+          type: {
+            react: ["react"],
+          },
+          value: {
+            react: ["react"],
+          },
+        },
+        groups: ["react", ...SORT_IMPORTS_GROUPS],
+        "ignore-case": true,
+        "newlines-between": "ignore",
+        type: "natural",
+      },
+    ],
+    "perfectionist/sort-jsx-props": [
+      "error",
+      {
+        "custom-groups": {
+          callback: "on*",
+          reservedProps: ["children", "dangerouslySetInnerHTML", "key", "ref"], // Reserved props from: https://github.com/jsx-eslint/eslint-plugin-react/blob/master/lib/rules/jsx-sort-props.js#L40C12-L40C12
+        },
+        groups: ["reservedProps", "unknown", "callback"],
+        "ignore-case": true,
+        type: "natural",
+      },
+    ],
   },
   settings: {
     react: {